Beginnings of project
This is a big commit that includes the init of the project, creation of
the StateOrTerritories resources and LoanRecord models.

This includes some primitive tests for hitting the index page and
clicking a state/territory

Some basic seeding of the data as well.

In summary this is the barebones of a rails project plus some basic
beginnings of more https://www.youtube.com/watch?v=lI-BMDnti4c